\name{local.block.vector-class}
\alias{local.block.vector-class}
\title{local.block.vector class}
\description{
  The local.block.vector class allows users to create
  distributed vectors that are stored in an environment in
  the calling R session. Please note that this class is
  mostly for testing purposes and provides an set-up that
  is easy to debug in.
}

