\name{rand.port}
\alias{rand.port}
\title{Creat a port number for transferring things over zeromq channels.}
\usage{
rand.port(start = 50000, end = 5e+05)
}
\arguments{
  \item{start}{The starting port}

  \item{end}{The ending port}
}
\description{
Creat a port number for transferring things over zeromq
channels.
}

